separate to classes, optimize refocusing
