+++ package.json
@@ -13,6 +13,7 @@
     "babel-loader": "^7.1.4",
     "babel-plugin-transform-runtime": "^6.23.0",
     "babel-preset-env": "^1.7.0",
+    "prettier": "^1.14.2",
     "webpack": "^4.13.0",
     "webpack-cli": "^3.0.8"
   }
